+++ src/app.module.ts
@@ -14,6 +14,7 @@ import { DatabaseBackupModule } from './modules/database-backup/database-backup.
 import { ComposeModule } from './modules/compose/compose.module';
 import { ComposeMountModule } from './modules/compose-mount/compose-mount.module';
 import { InstanceBackupModule } from './modules/instance-backup/instance-backup.module';
+import { InstanceUpdateModule } from './modules/instance-update/instance-update.module';
 import { InvitationModule } from './modules/invitation/invitation.module';
 import { DatabaseMountModule } from './modules/database-mount/database-mount.module';
 import { DatabaseModule } from './modules/database/database.module';
@@ -61,6 +62,7 @@ import { VolumeBackupModule } from './modules/volume-backup/volume-backup.module
     ComposeMountModule,
     InvitationModule,
     InstanceBackupModule,
+    InstanceUpdateModule,
     RegistryModule,
     ApplicationModule,
     ProxyModule,