+++ src/modules/registry/provision-self-hosted/provision-self-hosted.controller.ts
@@ -1,8 +1,11 @@
-import { Controller, Post, UseGuards } from '@nestjs/common';
+import { Body, Controller, Post, UseGuards } from '@nestjs/common';
 import { JwtAuthGuard } from '../../auth/jwt-auth.guard';
 import { Roles } from '../../auth/roles.decorator';
 import { RolesGuard } from '../../auth/roles.guard';
-import { ProvisionSelfHostedResponseDto } from './provision-self-hosted.dto';
+import {
+  ProvisionSelfHostedDto,
+  ProvisionSelfHostedResponseDto,
+} from './provision-self-hosted.dto';
 import { ProvisionSelfHostedService } from './provision-self-hosted.service';
 
 @Controller('registries')
@@ -12,7 +15,9 @@ export class ProvisionSelfHostedController {
   constructor(private readonly service: ProvisionSelfHostedService) {}
 
   @Post('self-hosted')
-  provision(): Promise<ProvisionSelfHostedResponseDto> {
-    return this.service.execute();
+  provision(
+    @Body() dto: ProvisionSelfHostedDto,
+  ): Promise<ProvisionSelfHostedResponseDto> {
+    return this.service.execute(dto);
   }
 }